fixed additional missing pieces on the refactor
