+++ applications/rollout_metrics_noisy_ic.py
@@ -31,7 +31,7 @@ from credit.metrics import LatWeightedMetrics, LatWeightedMetricsClimatology
 from credit.forecast import load_forecasts
 from credit.distributed import distributed_model_wrapper, setup, get_rank_info
 from credit.models.checkpoint import load_model_state, load_state_dict_error_handler
-from credit.postblock import GlobalMassFixer, GlobalWaterFixer, GlobalEnergyFixer
+from credit.postblock.gen1 import GlobalMassFixer, GlobalWaterFixer, GlobalEnergyFixer
 from credit.parser import credit_main_parser, predict_data_check
 from credit.datasets.era5_multistep_batcher import Predict_Dataset_Batcher
 from credit.datasets.load_dataset_and_dataloader import BatchForecastLenDataLoader
